# [.NET] Improved parsing time