--- a/dotnet/Gherkin/GherkinLanguageConstants.cs
+++ b/dotnet/Gherkin/GherkinLanguageConstants.cs
@@ -6,6 +6,7 @@ public static class GherkinLanguageConstants
     public const string COMMENT_PREFIX = "#";
     public const string TITLE_KEYWORD_SEPARATOR = ":";
     public const string TABLE_CELL_SEPARATOR = "|";
+    public const char TABLE_CELL_SEPARATOR_CHAR = '|';
     public const char TABLE_CELL_ESCAPE_CHAR = '\\';
     public const char TABLE_CELL_NEWLINE_ESCAPE = 'n';
     public const string DOCSTRING_SEPARATOR = "\"\"\"";