@obligaron 2024-12-16
No problem. 🙂 
Here are the results without the delegate optimization (second commit reverted):

| Method      | Runtime              | FeatureFile       | Mean        | Error     | StdDev    | Gen0     | Gen1     | Gen2    | Allocated  |
|------------ |--------------------- |------------------ |------------:|----------:|----------:|---------:|---------:|--------:|-----------:|
| **Parser**      | **.NET 8.0**             | **tags.feature**      |   **425.46 μs** |  **8.187 μs** | **10.645 μs** |  **35.1563** |  **35.1563** | **35.1563** |  **531.46 KB** |
| ParserReuse | .NET 8.0             | tags.feature      |    23.71 μs |  0.143 μs |  0.127 μs |   4.9744 |   0.1831 |       - |   81.66 KB |
| Parser      | .NET Framework 4.8.1 | tags.feature      |   839.09 μs |  1.712 μs |  1.602 μs |  76.1719 |  40.0391 | 35.1563 |  553.49 KB |
| ParserReuse | .NET Framework 4.8.1 | tags.feature      |    30.64 μs |  0.350 μs |  0.292 μs |  13.9771 |   0.4883 |       - |   86.04 KB |
| **Parser**      | **.NET 8.0**             | **very_long.feature** | **1,245.31 μs** | **16.312 μs** | **13.621 μs** | **210.9375** | **175.7813** | **35.1563** | **3428.37 KB** |
| ParserReuse | .NET 8.0             | very_long.feature |   831.35 μs | 16.616 μs | 17.063 μs | 181.6406 |  90.8203 |       - | 2973.31 KB |
| Parser      | .NET Framework 4.8.1 | very_long.feature | 2,217.68 μs | 43.897 μs | 41.061 μs | 550.7813 | 250.0000 | 35.1563 | 3493.88 KB |
| ParserReuse | .NET Framework 4.8.1 | very_long.feature | 1,223.84 μs |  8.885 μs |  6.937 μs | 490.2344 | 187.5000 |       - | 3015.27 KB |

Without it, `very_long.feature` with reuse decreases by 47% for .NET 8 and 35% for .NET Framework 4.8. Allocations almost double.

Regarding error handling:
Most methods are generated, the changes in the `gherking-csharp.razor` file adds only [22 lines](https://github.com/cucumber/gherkin/pull/336/files#diff-7b86eeddab5de8cb5c6b79db46f9502143af7605b05fa2c7ab4be05f0205787c).