@gasparnagy 2024-12-18
@obligaron Thx for the measure. Yes, this is significant enough to keep.