@luke-hill 2025-01-06
When I see 30-40% improvements in performance I'm always keen. Is this something that exists in a diff form in other flavours. Is there a location we could look to improve in other flavours too? cc/ @gasparnagy