@obligaron 2025-01-06
What is likely to transfer is how this PR and #344 try to minimise string creation/operations in `GherkinLine`.

Another thing might be the AstNode subitem handling.

Other optimisations depend more on the language/runtime being used.
If they are more like .NET (JIT and GC) more optimisations can be transferred. For example, the Java implementation could also change the error handling as is done in this PR.

Hope this helps a bit. 🙂