@luke-hill 2025-01-07
Thanks @obligaron made https://github.com/cucumber/gherkin/issues/351